:recycle: Extract issue list filtering to IssueFilterService
