+++ issuepoker-backend/src/main/java/de/muenchen/issuepoker/services/IssueService.java
@@ -6,17 +6,13 @@ import de.muenchen.issuepoker.common.ConflictException;
 import de.muenchen.issuepoker.common.NotFoundException;
 import de.muenchen.issuepoker.entities.issue.Issue;
 import de.muenchen.issuepoker.entities.issue.IssueKey;
-import de.muenchen.issuepoker.entities.vote.Vote;
 import de.muenchen.issuepoker.entities.issue.filter.FilterDTO;
 import de.muenchen.issuepoker.entities.issue.filter.FilterOptionsDTO;
 import de.muenchen.issuepoker.entities.issue.request.IssueRequestUpdateDTO;
+import de.muenchen.issuepoker.entities.vote.Vote;
 import de.muenchen.issuepoker.repositories.IssueRepository;
 import de.muenchen.issuepoker.security.Authorities;
-import jakarta.persistence.criteria.CriteriaBuilder;
-import jakarta.persistence.criteria.Predicate;
-import jakarta.persistence.criteria.Root;
 import java.util.List;
-import java.util.Locale;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.dao.DataIntegrityViolationException;
@@ -30,6 +26,7 @@ import org.springframework.stereotype.Service;
 @RequiredArgsConstructor
 public class IssueService {
     private final IssueRepository issueRepository;
+    private final IssueFilterService issueFilterService;
 
     @PreAuthorize(Authorities.IS_USER)
     public Issue getIssue(final IssueKey issueKey) {
@@ -42,53 +39,8 @@ public class IssueService {
     public Page<Issue> getIssueList(final Pageable pageRequest, final FilterDTO filter) {
         log.info("Get all Issues for page {} filtered by {}", pageRequest, filter);
         return issueRepository.findAll(
-                (root, query, criteriaBuilder) -> criteriaBuilder.and(
-                        filterSearch(filter.search(), root, criteriaBuilder),
-                        filterOwner(filter.owners(), root, criteriaBuilder),
-                        filterRepository(filter.repositories(), root, criteriaBuilder),
-                        filterVoted(filter.voted(), root, criteriaBuilder),
-                        filterResulted(filter.resulted(), root, criteriaBuilder)),
+                (root, query, criteriaBuilder) -> issueFilterService.filterIssues(filter, root, criteriaBuilder),
                 pageRequest);
-    }
-
-    private Predicate filterSearch(final String search, final Root<Issue> root, final CriteriaBuilder criteriaBuilder) {
-        if (search == null || search.isEmpty()) {
-            return criteriaBuilder.conjunction();
-        }
-        final String likePattern = "%" + search.toLowerCase(Locale.ROOT) + "%";
-        return criteriaBuilder.or(
-                criteriaBuilder.like(criteriaBuilder.lower(root.get("title")), likePattern),
-                criteriaBuilder.like(criteriaBuilder.lower(root.get("description")), likePattern));
-    }
-
-    private Predicate filterOwner(final List<String> owners, final Root<Issue> root, final CriteriaBuilder criteriaBuilder) {
-        return (owners != null && !owners.isEmpty())
-                ? root.get("owner").in(owners)
-                : criteriaBuilder.conjunction();
-    }
-
-    private Predicate filterRepository(final List<String> repositories, final Root<Issue> root, final CriteriaBuilder criteriaBuilder) {
-        return (repositories != null && !repositories.isEmpty())
-                ? root.get("repository").in(repositories)
-                : criteriaBuilder.conjunction();
-    }
-
-    private Predicate filterVoted(final Boolean voted, final Root<Issue> root, final CriteriaBuilder criteriaBuilder) {
-        if (voted == null) {
-            return criteriaBuilder.conjunction();
-        }
-        return voted
-                ? criteriaBuilder.isNotEmpty(root.get("votes"))
-                : criteriaBuilder.isEmpty(root.get("votes"));
-    }
-
-    private Predicate filterResulted(final Boolean resulted, final Root<Issue> root, final CriteriaBuilder criteriaBuilder) {
-        if (resulted == null) {
-            return criteriaBuilder.conjunction();
-        }
-        return resulted
-                ? criteriaBuilder.isNotNull(root.get("voteResult"))
-                : criteriaBuilder.isNull(root.get("voteResult"));
     }
 
     @PreAuthorize(Authorities.IS_ADMIN)
